requirements.txt
torch==1.12.1
git+https://github.com/discordianfish/LAVIS#egg=lavis
pillow
torchvision
transformers==4.15.0
timm==0.4.12
fairscale==0.4.4
datasets
